Extract collection items to a separate widget
